feat: refactor auth module
