+++ src/index.ts
@@ -10,4 +10,4 @@ export * from './wallet';
 export * from './events';
 
 export * from './modules';
-export * from './permits';
+export * from './auth'
